+++ modules/admin/src/ngrest/base/ActiveWindowInterface.php
@@ -12,6 +12,10 @@ interface ActiveWindowInterface
 {
     public function setItemId($id);
 
+    public function setConfigHash($hash);
+    
+    public function setActiveWindowHash($hash);
+    
     public function getItemId();
 
     public function index();